requirements.txt
torch>=2.0
torchvision>=0.15
timm>=1.0.20
transformers>=4.56.0
albumentations>=1.4.0
tensorboard
tqdm
matplotlib
numpy
Pillow
opencv-python
PyYAML
scipy
